Changes in route, now the role management page would ask to be at least a supervisor to see the page
